refactor: query/sites/open 路由结构优化
